+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/customer/application/CustomerRegionService.java
@@ -2,7 +2,7 @@ package com.breaditnow.customer.customer.application;
 
 import com.breaditnow.customer.customer.application.request.RegionUpdateRequest;
 import com.breaditnow.customer.customer.domain.port.SaveCustomerRegionPort;
-import com.breaditnow.customer.region.application.port.LoadRegionPort;
+import com.breaditnow.customer.region.domain.port.LoadRegionPort;
 import com.breaditnow.domain.global.exception.DomainErrorCode;
 import com.breaditnow.domain.global.exception.DomainException;
 import lombok.RequiredArgsConstructor;
