+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/region/infrastructure/RegionAdapter.java
@@ -1,7 +1,7 @@
 package com.breaditnow.customer.region.infrastructure;
 
-import com.breaditnow.customer.region.application.port.LoadRegionPort;
-import com.breaditnow.customer.region.core.Region;
+import com.breaditnow.customer.region.domain.port.LoadRegionPort;
+import com.breaditnow.customer.region.domain.Region;
 import com.breaditnow.customer.region.infrastructure.entity.RegionEntity;
 import com.breaditnow.customer.region.infrastructure.jpa.JpaRegionRepository;
 import com.breaditnow.customer.region.presentation.res.GugunResponse;
